+++ docs/changelog.md
@@ -4,6 +4,7 @@
 
 ### Enhancements
 
+* [#3](https://github.com/netboxlabs/nbl-netbox-branching/issues/3) - Require review & acknowledgment of conflicts before syncing or merging a branch
 * [#4](https://github.com/netboxlabs/nbl-netbox-branching/issues/4) - Include a three-way diff summary in the REST API representation of a modified object
 
 ### Bug Fixes